api/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
bcrypt==4.1.2
PyJWT==2.10.1
python-dotenv==1.1.1
supabase==2.18.1
postgrest==1.1.1
python-dateutil==2.9.0.post0
Werkzeug==3.0.1
APScheduler==3.10.4
